+++ src/command_modules/azure-cli-consumption/azure/cli/command_modules/consumption/_help.py
@@ -17,12 +17,21 @@ helps['consumption reservations'] = """
 
 helps['consumption reservations summaries'] = """
     type: group
-    short-summary: Manage reservations summaries for daily or monthly.
+    short-summary: List reservations summaries.
+"""
+helps['consumption reservations summaries list'] = """
+    type: command
+    short-summary: List reservations summaries for daily or monthly by reservation order Id and optionally by reservation id.
 """
 
 helps['consumption reservations details'] = """
     type: group
-    short-summary: Manage reservations details.
+    short-summary: List reservations details.
+"""
+
+helps['consumption reservations details list'] = """
+    type: command
+    short-summary: List reservations details by reservation order id and optionally by reservation id
 """
 
 helps['consumption usage'] = """
@@ -33,4 +42,14 @@ helps['consumption usage'] = """
 helps['consumption usage list'] = """
     type: command
     short-summary: Show the details of Azure resource consumption, either as an invoice or within a billing period.
+"""
+
+helps['consumption pricesheet'] = """
+    type: group
+    short-summary: Inspect the price sheet by Azure subscription within current or specified billing period.
+"""
+
+helps['consumption pricesheet show'] = """
+    type: command
+    short-summary: Show the price sheet for Azure subscription within a current billing period or specified billing period
 """