+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/_help.py
@@ -21,6 +21,8 @@ helps['container create'] = """
           text: az container create -g MyResourceGroup --name mywinapp --image winappimage:latest --os-type Windows --cpu 2 --memory 3.5
         - name: Create a container in a container group with public IP address and ports.
           text: az container create -g MyResourceGroup --name myalpine --image alpine:latest --ip-address public --ports 80 443
+        - name: Create a container in a container group with public IP address, ports and DNS name label.
+          text: az container create -g MyResourceGroup --name myalpine --image alpine:latest --ports 80 443 --dns-name-label contoso
         - name: Create a container in a container group that invokes a script upon start.
           text: az container create -g MyResourceGroup --name myalpine --image alpine:latest --command-line "/bin/sh -c '/path to/myscript.sh'"
         - name: Create a container in a container group that runs a command and stop the container afterwards.
@@ -52,5 +54,10 @@ helps['container show'] = """
 
 helps['container logs'] = """
     type: command
-    short-summary: Examine the logs for a container group.
+    short-summary: Examine the logs for a container in a container group.
+"""
+
+helps['container attach'] = """
+    type: command
+    short-summary: Attach local standard output and error streams to a container in a container group.
 """