+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/tests/latest/test_container_commands.py
@@ -9,7 +9,7 @@ from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer
 
 
 class AzureContainerInstanceScenarioTest(ScenarioTest):
-    # Test create container with image, os type, ip address type, port, cpu,
+    # Test create container with image, os type, ip address type, port, dns name label, cpu,
     # memory, command line and environment variables specified.
     @ResourceGroupPreparer()
     def test_container_create(self, resource_group, resource_group_location):
@@ -17,6 +17,8 @@ class AzureContainerInstanceScenarioTest(ScenarioTest):
         image = 'alpine:latest'
         os_type = 'Linux'
         ip_address_type = 'Public'
+        dns_name_label = container_group_name
+        fqdn = '{}.{}.azurecontainer.io'.format(container_group_name, resource_group_location)
         port1 = 8000
         port2 = 8001
         ports = '{} {}'.format(port1, port2)
@@ -25,12 +27,16 @@ class AzureContainerInstanceScenarioTest(ScenarioTest):
         command = '"/bin/sh -c \'while true; do echo hello; sleep 20; done\'"'
         env = 'KEY1=VALUE1 KEY2=FOO=BAR='
         restart_policy = 'Never'
+        secrets = 'secret1=superawesomesecret secret2="nothing to see"'
+        secret_path = '/s'
 
         self.kwargs.update({
             'container_group_name': container_group_name,
             'resource_group_location': resource_group_location,
             'image': image,
             'os_type': os_type,
+            'dns_name_label': dns_name_label,
+            'fqdn': fqdn,
             'ip_address_type': ip_address_type,
             'port1': port1,
             'port2': port2,
@@ -39,18 +45,23 @@ class AzureContainerInstanceScenarioTest(ScenarioTest):
             'memory': memory,
             'command': command,
             'env': env,
-            'restart_policy': restart_policy
+            'restart_policy': restart_policy,
+            'secrets': secrets,
+            'secrets_mount_path': secret_path
         })
 
         # Test create
         self.cmd('container create -g {rg} -n {container_group_name} --image {image} --os-type {os_type} '
-                 '--ip-address {ip_address_type} --ports {ports} --cpu {cpu} --memory {memory} '
-                 '--command-line {command} -e {env} --restart-policy {restart_policy}',
+                 '--ip-address {ip_address_type} --dns-name-label {dns_name_label} --ports {ports} --cpu {cpu} --memory {memory} '
+                 '--command-line {command} -e {env} --restart-policy {restart_policy} '
+                 '--secrets {secrets} --secrets-mount-path {secrets_mount_path}',
                  checks=[self.check('name', '{container_group_name}'),
                          self.check('location', '{resource_group_location}'),
                          self.check('provisioningState', 'Creating'),
                          self.check('osType', '{os_type}'),
                          self.check('restartPolicy', '{restart_policy}'),
+                         self.check('ipAddress.dnsNameLabel', '{container_group_name}'),
+                         self.check('ipAddress.fqdn', '{fqdn}'),
                          self.exists('ipAddress.ip'),
                          self.exists('ipAddress.ports'),
                          self.check('ipAddress.ports[0].port', '{port1}'),
@@ -59,7 +70,9 @@ class AzureContainerInstanceScenarioTest(ScenarioTest):
                          self.exists('containers[0].command'),
                          self.exists('containers[0].environmentVariables'),
                          self.check('containers[0].resources.requests.cpu', cpu),
-                         self.check('containers[0].resources.requests.memoryInGb', memory)])
+                         self.check('containers[0].resources.requests.memoryInGb', memory),
+                         self.exists('volumes'),
+                         self.check('volumes[0].secret', {})])
 
         # Wait for container to be provisioned
         time.sleep(30)