+++ src/command_modules/azure-cli-interactive/azclishell/app.py
@@ -631,7 +631,7 @@ class AzInteractiveShell(object):
                 self.config.set_feedback('yes')
                 self.user_feedback = False
 
-            azure_folder = self.config.config_dir
+            azure_folder = get_config_dir()
             if not os.path.exists(azure_folder):
                 os.makedirs(azure_folder)
             ACCOUNT.load(os.path.join(azure_folder, 'azureProfile.json'))