+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/commands.py
@@ -477,7 +477,7 @@ def load_command_table(self, _):
     # endregion
 
     # region NetworkWatchers
-    with self.command_group('network watcher', network_watcher_sdk, min_api='2016-09-01') as g:
+    with self.command_group('network watcher', network_watcher_sdk, client_factory=cf_network_watcher, min_api='2016-09-01') as g:
         g.custom_command('configure', 'configure_network_watcher')
         g.command('list', 'list_all')
         g.custom_command('test-ip-flow', 'check_nw_ip_flow', client_factory=cf_network_watcher)
@@ -619,7 +619,7 @@ def load_command_table(self, _):
         g.command('list-advertised-routes', 'get_advertised_routes')
         g.command('list-learned-routes', 'get_learned_routes')
 
-    with self.command_group('network vnet-gateway vpn-client', network_vgw_sdk) as g:
+    with self.command_group('network vnet-gateway vpn-client', network_vgw_sdk, client_factory=cf_virtual_network_gateways) as g:
         g.custom_command('generate', 'generate_vpn_client')
         g.command('show-url', 'get_vpn_profile_package_url', min_api='2017-08-01')