+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/tests/latest/test_storage_account_scenarios.py
@@ -170,7 +170,7 @@ class StorageAccountTests(StorageScenarioMixin, ScenarioTest):
     @StorageAccountPreparer()
     def test_metrics_operations(self, resource_group, storage_account_info):
         self.storage_cmd('storage metrics show', storage_account_info) \
-            .assert_with_checks(JMESPathCheck('file.hour.enabled', False),
+            .assert_with_checks(JMESPathCheck('file.hour.enabled', True),
                                 JMESPathCheck('file.minute.enabled', False))
 
         self.storage_cmd('storage metrics update --services f --api true --hour true --minute true --retention 1 ',