+++ src/command_modules/azure-cli-interactive/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.3.16
+++++++
+* Fix issue where user is prompted to login when using interactive mode in Cloud Shell.
+* Fixed regression with missing parameter completions.
+
 0.3.15
 ++++++
 * Fixed issue where command option completions no longer appeared.