+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -3,6 +3,19 @@
 Release History
 ===============
 
+2.0.24
+++++++
+* `network vnet-gateway vpn-client generate`: Fix missing client issue.
+
+2.0.23
+++++++
+* `network public-ip create`: Fix `--tags` option.
+* `network lb create`: Fix `--tags` option.
+* `network local-gateway create`: Fix `--tags` option.
+* `network nic create`: Fix `--tags` option.
+* `network vnet-gateway create`: Fix `--tags` option.
+* `network vpn-connection create`: Fix `--tags` option.
+
 2.0.22
 ++++++
 * `application-gateway create`: `--cert-password` protected using secureString.